% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esir.R
\name{elptstd}
\alias{elptstd}
\title{Scatter Matrix Estimation for Elliptical Distributions by the variant of Hettmansperger-Randles estimate}
\usage{
elptstd(X, epsilon = 0.005, maxiter = 1000)
}
\arguments{
\item{X}{A numeric matrix of observations with dimension \code{n x p}.}

\item{epsilon}{A positive numeric scalar specifying the convergence
tolerance. Defaults to \code{5e-3}.}

\item{maxiter}{A positive integer specifying the maximum number of
iterations. Defaults to \code{1000}.}
}
\value{
A \code{p x p} symmetric positive definite scatter matrix.
}
\description{
Estimates the scatter matrix of the covariate matrix by iteratively updating eigenvalues 
of the spatial Kendall's tau covariance matrix until convergence. See Li et al. (2026).
}
