% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esir.R
\name{matpower}
\alias{matpower}
\title{Matrix Power}
\usage{
matpower(mat, power)
}
\arguments{
\item{mat}{A \code{p x p} symmetric positive definite matrix.}

\item{power}{A numeric scalar specifying the power to raise \code{mat} to.}
}
\value{
A \code{p x p} matrix equal to \eqn{A^{\text{power}}}.
}
\description{
Computes an arbitrary real power of a symmetric positive definite matrix
via eigendecomposition.
}
\keyword{internal}
