% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sike-package.R
\docType{package}
\name{sike-package}
\alias{sike-package}
\title{sike: Sliced Inverse Kendall's tau Estimation}
\description{
R package for Sliced Inverse Kendall's tau Estimation (SIKE) for sufficient dimension reduction with heavy-tailed covariates. See the published paper High-dimensional Sliced Inverse Kendall’s tau Estimation for Heavy-tailed Data.
}
